backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
bcrypt<5.0.0
python-multipart
redis
celery[redis]
requests
beautifulsoup4
prometheus-client
starlette-exporter
python-json-logger
elasticsearch
logstash-formatter
pydantic-settings
email-validator
